requirements.txt
Flask==2.3.2
pymongo==4.3.3
python-dotenv==0.21.0
werkzeug==2.3.3
